+++ ele-autotesting/packages/web/src/App.vue
@@ -59,6 +59,7 @@
           </template>
         </ActionButtonUI>
         <ThemeToggleUI />
+        <span class="ds-vrule hidden sm:inline-block" aria-hidden="true"></span>
         <ActionButtonUI text="功能提示词" @click="openTemplateManager">
           <template #icon>
             <svg xmlns="http://www.w3.org/2000/svg" class="h-4 w-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round">